Rename pkg/network/vmicontroller to pkg/network/controllers
Currently, the `vmicontroller` package contains logic calculating
the VMI interfaces status.

Rename the package to `controllers` in preparation for the move of the
network VM controller in a following commit so its name would reflect
its future content.

Rename status.go and status_test.go to vmi.go and vmi_test.go in order
to better reflect their content.
Rename the `UpdateStatus` function to `UpdateVMIStatus`.
