+++ pkg/network/controllers/BUILD.bazel
@@ -2,8 +2,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["status.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/network/vmicontroller",
+    srcs = ["vmi.go"],
+    importpath = "kubevirt.io/kubevirt/pkg/network/controllers",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/network/multus:go_default_library",
@@ -18,8 +18,8 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "status_test.go",
-        "vmicontroller_suite_test.go",
+        "controllers_suite_test.go",
+        "vmi_test.go",
     ],
     deps = [
         ":go_default_library",
